+++ src/components/organisms/Form/FormTextField/FormTextField.tsx
@@ -1,6 +1,6 @@
 import React, { FC, ChangeEvent } from 'react';
 import TextField, { ITextFieldProps } from '../../../molecules/TextField/TextField';
-import { useFieldContext } from '../../../hooks/useForm';
+import { useFieldContext } from '../useForm';
 
 interface IFormTextFieldProps extends ITextFieldProps {
   name: string;
