+++ src/components/organisms/Form/index.tsx
@@ -1,14 +1,13 @@
 import React, { FC, HTMLAttributes } from 'react';
 
-import { IUseFormProps } from '../../hooks/useForm/useForm';
-
+import { IUseFormProps } from './useForm';
 import FormButton from './FormButton/FormButton';
 import FormTextField from './FormTextField/FormTextField';
 import FormCheckboxField from './FormCheckboxField/FormCheckboxField';
 import FormRadioField from './FormRadioField/FormRadioField';
 import FormDropdownField from './FormDropdownField/FormDropdownField';
 import FormDropdownFilteredField from './FormDropdownFilteredField/FormDropdownFilteredField';
-import Form from './Form/Form';
+import FormComponent from './Form/Form';
 
 interface IFormStatic {
   Button: typeof FormButton;
@@ -19,13 +18,15 @@ interface IFormStatic {
   DropdownFilteredField: typeof FormDropdownFilteredField;
 }
 
-const FormWrapper: IFormStatic & FC<HTMLAttributes<HTMLFormElement> & IUseFormProps> = props => <Form {...props} />;
+export const Form: IFormStatic & FC<HTMLAttributes<HTMLFormElement> & IUseFormProps> = props => (
+  <FormComponent {...props} />
+);
 
-FormWrapper.Button = FormButton;
-FormWrapper.TextField = FormTextField;
-FormWrapper.CheckboxField = FormCheckboxField;
-FormWrapper.RadioField = FormRadioField;
-FormWrapper.DropdownField = FormDropdownField;
-FormWrapper.DropdownFilteredField = FormDropdownFilteredField;
+Form.Button = FormButton;
+Form.TextField = FormTextField;
+Form.CheckboxField = FormCheckboxField;
+Form.RadioField = FormRadioField;
+Form.DropdownField = FormDropdownField;
+Form.DropdownFilteredField = FormDropdownFilteredField;
 
-export default FormWrapper;
+export * from './useForm';
